requirements.txt
sentence-transformers
faiss-cpu
numpy
scikit-learn
pypdf
vectorstore
groq
python-dotenv